Add files through upload
"This repository contains the initial upload of files for a Progressive Web App (PWA) project designed for a pharmacy system. The files were originally obtained from a partially completed project and are now uploaded here as my first repository on GitHub. The goal is to further develop and integrate this PWA, with plans to deploy it on Netlify."